requirements.txt
python-telegram-bot==20.7
openai==1.3.7
okama==1.2.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
plotly==5.15.0
kaleido==0.2.1
